+++ database/migrations/072_dashboard_suporte_zabbix.sql
@@ -1,0 +1,1 @@
+UPDATE auth_perfis SET dashboard_principal = 'infraestrutura.monitoramento_zabbix' WHERE codigo = 'SUPORTE';